better code for dice rolling, yet to start on the 3rd assignment
