+++ Simulation & Modelling/Assignments/Server Queuing/infile.txt
@@ -1,1 +1,1 @@
-1.0 0.5 467
+1.0 0.5 1000
